kconfig: Parse help text in CONFIG_XXX.
Usually, the help keyword starts with \t or \s, and the help text
starts with the same room.

Most of them starts with \t\s\s, but there are some irregulars
without the double spaces.

Also, the help text sometimes contains a newline.  Thus, we need to
parse newline and the _next_ line.  If the next line does not start
with the spaces, we can give up parseing the text.

Then, we need to undo readline() by fseek() and set True to break_pasre,
otherwise, the next line will be parsed within Config.pasre(), instead
of the parent Menut.parse().

When bailing out, we need to update the parent.lines[-1], so that the
parent can track the correct line number.

Probably we can do much simpler like parsing global keywords and bail
out.. but let's do real parsing first and get back here when needed.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>